move tc errors to separate module and remove redundant ones
