+++ library/Core/src/Juvix/Core/IR/Typechecker.hs
@@ -6,12 +6,14 @@
 module Juvix.Core.IR.Typechecker
   ( module Juvix.Core.IR.Typechecker,
     module Typed,
+    module Error,
     module Env,
   )
 where
 
 import Juvix.Core.IR.Typechecker.Env as Env
 import Juvix.Core.IR.Typechecker.Types as Typed
+import Juvix.Core.IR.Typechecker.Error as Error
 import qualified Juvix.Core.IR.Types as IR
 import qualified Juvix.Core.IR.Types.Base as IR
 import qualified Juvix.Core.IR.Types.Globals as IR
